reorganize to cleanup typedoc output and clarify API vs. impl
