+++ Gruntfile.js
@@ -26,10 +26,10 @@ module.exports = function(grunt) {
           //reporters: ['pretty'],
           reporters: ['runner'],
           suites: [
-            //'dist/commonjs/es5/**/*.js',
-            //'!dist/commonjs/es5/**/BigList*.js',
-            //'dist/commonjs/es5/**/IonHash*.js',
-            'dist/commonjs/es5/**/FieldnameB*.js',
+            'dist/commonjs/es5/**/*.js',
+            '!dist/commonjs/es5/**/BigList*.js',
+            '!dist/commonjs/es5/**/IonHash*.js',
+            //'dist/commonjs/es5/**/FieldnameB*.js',
           ],
         },
       },
@@ -42,10 +42,11 @@ module.exports = function(grunt) {
           out: 'docs/api',
           module: 'commonjs',
           target: 'es5',
+          excludeExternals: true,
           excludePrivate: true,
           hideGenerator: true,
         },
-        src: 'src/**/*.ts',
+        src: 'src/*.ts',
       },
     },
   });
